Refactor test logger, update docs and naming conventions
- Updated copilot-instructions.md: clarified PowerShell usage and added rule against single character variable names.
- Refactored AuditLogTests: moved TestAuditLogger to its own file and made AuditLogTests partial.
- Improved naming in EnvironmentContext.cs by renaming lambda parameters for clarity.
