+++ src/Xcaciv.Command.Tests/AuditLogTests.cs
@@ -1,15 +1,13 @@
 using Xunit;
 using Xcaciv.Command;
-using Xcaciv.Command.Interface;
 using Xcaciv.Command.Tests.TestImpementations;
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Xunit.Abstractions;
 
 namespace Xcaciv.Command.Tests
 {
-    public class AuditLogTests
+    public partial class AuditLogTests
     {
         private ITestOutputHelper _testOutput;
 
@@ -213,70 +211,6 @@ namespace Xcaciv.Command.Tests
             Assert.Equal("SAY", logger.ExecutionLogs[0].CommandName);
             Assert.Equal("SAY", logger.ExecutionLogs[1].CommandName);
             Assert.Equal("SAY", logger.ExecutionLogs[2].CommandName);
-        }
-
-        /// <summary>
-        /// Test audit logger that captures all logs for testing
-        /// </summary>
-        private class TestAuditLogger : IAuditLogger
-        {
-            public List<CommandExecutionLog> ExecutionLogs { get; } = new();
-            public List<EnvironmentChangeLog> EnvironmentLogs { get; } = new();
-
-            public void LogCommandExecution(
-                string commandName,
-                string[] parameters,
-                DateTime executedAt,
-                TimeSpan duration,
-                bool success,
-                string? errorMessage = null)
-            {
-                ExecutionLogs.Add(new CommandExecutionLog
-                {
-                    CommandName = commandName,
-                    Parameters = parameters,
-                    ExecutedAt = executedAt,
-                    Duration = duration,
-                    Success = success,
-                    ErrorMessage = errorMessage
-                });
-            }
-
-            public void LogEnvironmentChange(
-                string variableName,
-                string? oldValue,
-                string? newValue,
-                string changedBy,
-                DateTime changedAt)
-            {
-                EnvironmentLogs.Add(new EnvironmentChangeLog
-                {
-                    VariableName = variableName,
-                    OldValue = oldValue,
-                    NewValue = newValue,
-                    ChangedBy = changedBy,
-                    ChangedAt = changedAt
-                });
-            }
-
-            public class CommandExecutionLog
-            {
-                public string CommandName { get; set; } = string.Empty;
-                public string[] Parameters { get; set; } = Array.Empty<string>();
-                public DateTime ExecutedAt { get; set; }
-                public TimeSpan Duration { get; set; }
-                public bool Success { get; set; }
-                public string? ErrorMessage { get; set; }
-            }
-
-            public class EnvironmentChangeLog
-            {
-                public string VariableName { get; set; } = string.Empty;
-                public string? OldValue { get; set; }
-                public string? NewValue { get; set; }
-                public string ChangedBy { get; set; } = string.Empty;
-                public DateTime ChangedAt { get; set; }
-            }
         }
     }
 }
